Squashed 'vision/' changes from ed1c26e02e..9328f3d1c4
9328f3d1c4 Deleted yml files
c36e096097 Corrected image names, added zynq emulation support
febbbda89e updated xf::mat to xf::cv::mat in api-reference.rst
9ea6dc8b4a Deleted all yml files
ebfca02591 Changed badpixelcorrection size to hd image
1af026bc85 Corrected top function name in tcl files
fe134fbfa0 Added depth value for cosim in colordetect and meanstddev
60d75c82a8 Corrected opt-flow i/p images in Makefiles
18886bfd5a Removed nested dataflow in warptransform hpp
523bcfce4d Updated canny Makefile, warptransform tb and gaussiandiff accel
bc84d1f57d Updated Makefiles, tcl files and input images
dbea5126a4 Updated L3 x+ml readme
9b0dace7d3 Corrected i/p images of 2 L1 functions
57ef7d2140 Corrected input image in L1 BPC function
065f31eb05 Corrected top function name in some L1 functions
685215978c Updated Clk frequency flag in Makefiles

git-subtree-dir: vision
git-subtree-split: 9328f3d1c4c3847b54ab399bc04224650bef11f1

+++ L2/examples/reduce/xf_reduce_tb.cpp
@@ -35,10 +35,10 @@ int main(int argc, char** argv) {
     }
 
 #if DIM
-    dst_hls.create(in_img.rows, 1, CV_8UC1);
+    dst_hls.create(in_img.rows, 1, CV_32SC1);
     ocv_ref.create(in_img.rows, 1, CV_8UC1);
 #else
-    dst_hls.create(1, in_img.cols, CV_8UC1);
+    dst_hls.create(1, in_img.cols, CV_32SC1);
     ocv_ref.create(1, in_img.cols, CV_8UC1);
 #endif
